Updated Loginout button to show relevant msg depending on logged in state.
Added a commented out window flag to main window. This to remove the top bar in the end product. Had to import QtCore from PyQT5 in order to do this.

Imported time module to track time being logged in, with the aim to log out inactiv user.

Created connect functions to radiobutton A/B. Currently it only prints to terminal, but it might be relevant to change out text in coffee labels due to this.